# docs/openapi-v1.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/admin/analytics/overview": {
      "get": {
        "operationId": "AdminAnalyticsController_overview",
        "summary": "Dashboard summary analytics",
        "parameters": [
          {
            "name": "range",
            "required": false,
            "in": "query",
            "description": "Date range preset. Aliases: last7Days → last7, last30Days → last30.",
            "schema": {
              "default": "last30",
              "enum": [
                "today",
                "last7",
                "last7Days",
                "last30",
                "last30Days",
                "thisMonth",
                "custom"
              ],
              "type": "string"
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "description": "ISO date (custom range start)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "description": "ISO date (custom range end)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dateFrom",
            "required": false,
            "in": "query",
            "description": "Alias for `from` (custom range start)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dateTo",
            "required": false,
            "in": "query",
            "description": "Alias for `to` (custom range end)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Overview analytics",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "tags": [
          "admin-analytics"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/admin/analytics/charts": {
      "get": {
        "operationId": "AdminAnalyticsController_charts",
        "summary": "Dashboard chart analytics",
        "parameters": [
          {
            "name": "range",
            "required": false,
            "in": "query",
            "description": "Date range preset. Aliases: last7Days → last7, last30Days → last30.",
            "schema": {
              "default": "last30",
              "enum": [
                "today",
                "last7",
                "last7Days",
                "last30",
                "last30Days",
                "thisMonth",
                "custom"
              ],
              "type": "string"
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "description": "ISO date (custom range start)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "description": "ISO date (custom range end)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dateFrom",
            "required": false,
            "in": "query",
            "description": "Alias for `from` (custom range start)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dateTo",
            "required": false,
            "in": "query",
            "description": "Alias for `to` (custom range end)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Chart analytics",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "tags": [
          "admin-analytics"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Tharaa Market API",
    "description": "Official **Version 1** REST API contract for Tharaa Market (Al Khafji).\n\n### Stability\nThis API is **frozen**. Breaking changes to URLs, request/response bodies, enums, pagination, auth, uploads, or error envelopes are not allowed in v1.\nFuture incompatible changes must ship as `/api/v2`.\n\n### Response envelope\nSuccess: `{ success: true, data, meta? }`\nError: `{ success: false, statusCode, message, errorCode?, details?, timestamp, path }`\n\n### Authentication\nUse `Authorization: Bearer <accessToken>`. Obtain tokens via `/auth/login`, `/auth/register`, `/auth/google`, or `/auth/refresh`.\nRefresh tokens rotate on every use.\n\n### Uploads\nMultipart form field name: `file`. Allowed MIME: jpeg, png, webp. Served under `/static/`.\n\n### Admin realtime (Socket.IO)\nNamespace `/admin` for staff dashboards. Authenticate with the JWT access token (`auth.token` or `Authorization: Bearer`).\nEvents: `admin:order_created`, `admin:order_updated`. Roles: ADMIN, MANAGER, EMPLOYEE.\nFull contract: `docs/ADMIN_INTEGRATION.md` (Realtime section). REST remains the source of truth.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Access JWT from auth endpoints"
      }
    },
    "schemas": {
      "ApiErrorDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false,
            "description": "Always false for errors"
          },
          "statusCode": {
            "type": "number",
            "example": 400,
            "description": "HTTP status code"
          },
          "message": {
            "type": "string",
            "example": "Validation failed",
            "description": "Human-readable error message"
          },
          "errorCode": {
            "type": "string",
            "example": "Bad Request",
            "description": "Optional machine-readable error code"
          },
          "details": {
            "type": "object",
            "description": "Optional extra details (e.g. validation messages)"
          },
          "timestamp": {
            "type": "string",
            "example": "2026-07-14T00:00:00.000Z",
            "description": "ISO-8601 timestamp when the error was produced"
          },
          "path": {
            "type": "string",
            "example": "/api/v1/products",
            "description": "Request path that produced the error"
          }
        },
        "required": [
          "success",
          "statusCode",
          "message",
          "timestamp",
          "path"
        ]
      },
      "ApiSuccessDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true,
            "description": "Always true for success"
          },
          "data": {
            "type": "object",
            "description": "Payload returned by the endpoint"
          },
          "meta": {
            "type": "object",
            "description": "Optional pagination or extra metadata",
            "additionalProperties": true
          }
        },
        "required": [
          "success",
          "data"
        ]
      },
...
    }
  }
}